+++ examples/web-demo/src/App.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import {browserIndexer, state, stringify} from '$lib/state/State';
+	import {browserIndexer, state, numRequests} from '$lib/state/State';
 	import JSONTree from 'svelte-json-tree';
 
 	function addLengthToFields(v: any): any {
@@ -19,13 +19,12 @@
 
 <progress value={($browserIndexer?.syncPercentage || 0) / 100} style="width:100%;" />
 
+<p>requests sent: {$numRequests}</p>
 <p>block processed: {$browserIndexer?.numBlocksProcessedSoFar.toLocaleString()}</p>
 <p>num events: {$browserIndexer?.nextStreamID.toLocaleString()}</p>
 
 {#if $state}
 	<JSONTree value={stateDisplayed} />
-{:else}{/if}
-
-<!-- <details>
-	{$state ? stringify($state) : $browserIndexer ? stringify($browserIndexer) : 'Please wait...'}
-</details> -->
+{:else if $browserIndexer}
+	<JSONTree value={$browserIndexer} />
+{/if}