+++ packages/ethereum-indexer-utils/CHANGELOG.md
@@ -1,5 +1,13 @@
 # ethereum-indexer
 
+## 0.3.3
+
+### Patch Changes
+
+- republish with new types
+- Updated dependencies
+  - ethereum-indexer@0.3.3
+
 ## 0.3.2
 
 ### Patch Changes